requirements.txt
cachelib==0.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cs50==5.1.0
Flask==1.1.2
Flask-Session==0.3.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
quickchart.io==0.1.0
requests==2.24.0
python-slugify==4.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
psycopg2
